Add delete icon hover styling and fix delete handler

+++ src/pages/List/List.js
@@ -100,11 +100,13 @@ const List = (props) => {
 
   const deleteItemHandler = async (e) => {
     try {
-      const response = await axios.delete(
-        `${BACKEND_URL}/lists/${listId}/${e.target.parentNode.id}`
-      );
-      console.log(response);
-      getList();
+      if (e.target.parentNode.id) {
+        const response = await axios.delete(
+          `${BACKEND_URL}/lists/${listId}/${e.target.parentNode.id}`
+        );
+        console.log(response);
+        getList();
+      }
     } catch (err) {
       console.log(`Error: ${err}`);
     }
